+++ omega_logger/__init__.py
@@ -5,7 +5,7 @@ from .validators import validator_map
 
 __author__ = 'Measurement Standards Laboratory of New Zealand'
 __copyright__ = '\xa9 2018 - 2022, ' + __author__
-__version__ = '0.4.0.dev0'
+__version__ = '0.4.0'
 
 _v = re.search(r'(\d+)\.(\d+)\.(\d+)[.-]?(.*)', __version__).groups()